fixed logic with access to project members
