+++ apps/models/app_types/jupyter.py
@@ -1,10 +1,30 @@
 from django.db import models
+from datetime import datetime, timedelta
+from django.db.models import Q
 
 from apps.models import AppInstanceManager, BaseAppInstance, Social
 
 
 class JupyterInstanceManager(AppInstanceManager):
     model_type = "jupyterinstance"
+    
+    def get_app_instances_of_project_filter(self, user, project, include_deleted=False, deleted_time_delta=None):
+        q = Q()
+
+        if not include_deleted:
+            if deleted_time_delta is None:
+                q &= ~Q(app_status__status="Deleted")
+            else:
+                time_threshold = datetime.now() - timedelta(minutes=deleted_time_delta)
+                q &= ~Q(app_status__status="Deleted") | Q(deleted_on__gte=time_threshold)
+
+        q &= Q(owner=user) | Q(
+            access__in=["project", "public", "private", "link"] if user.is_superuser else ["project", "public", "link"]
+        )
+        q &= Q(project=project)
+
+        return q
+   
 
 
 class JupyterInstance(BaseAppInstance):
