+++ apps/models/app_types/volume.py
@@ -10,21 +10,6 @@ from apps.models import AppInstanceManager, BaseAppInstance
 class VolumeInstanceManager(AppInstanceManager):
     model_type = "volumeinstance"
 
-    def get_app_instances_of_project_filter(self, user, project, include_deleted=False, deleted_time_delta=None):
-        q = Q()
-
-        if not include_deleted:
-            if deleted_time_delta is None:
-                q &= ~Q(app_status__status="Deleted")
-            else:
-                time_threshold = datetime.now() - timedelta(minutes=deleted_time_delta)
-                q &= ~Q(app_status__status="Deleted") | Q(deleted_on__gte=time_threshold)
-
-        q &= Q(owner=user)
-        q &= Q(project=project)
-
-        return q
-
 
 class VolumeInstance(BaseAppInstance):
     objects = VolumeInstanceManager()
